revert: select split into display mode
