+++ src/components/EarnFilterBar/components/EarnFilterSort.tsx
@@ -1,9 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { Select } from '../../core/form/Select/Select';
-import {
-  SelectDisplayMode,
-  SelectVariant,
-} from '../../core/form/Select/Select.types';
+import { SelectVariant } from '../../core/form/Select/Select.types';
 import { useEarnFiltering } from 'src/app/ui/earn/EarnFilteringContext';
 import { SortByEnum, SortByOptions } from 'src/app/ui/earn/types';
 
@@ -27,7 +24,6 @@ export const EarnFilterSort: React.FC = () => {
       variant={SelectVariant.Single}
       data-testid="earn-filter-sort-select"
       menuPlacementX="right"
-      displayMode={SelectDisplayMode.Menu}
     />
   );
 };
