+++ src/components/core/form/Select/Select.styles.ts
@@ -221,9 +221,3 @@ export const StyledSliderRangeContainer = styled(Box)(({ theme }) => ({
   justifyContent: 'space-between',
   width: '100%',
 }));
-
-export const StyledDrawerHeader = styled(Box)(({ theme }) => ({
-  height: 40,
-  overflow: 'hidden',
-  position: 'relative',
-}));
