+++ src/components/core/form/Select/Select.tsx
@@ -4,7 +4,6 @@ import {
   MultiSelectProps,
   SingleSelectProps,
   SliderSelectProps,
-  SelectDisplayMode,
 } from './Select.types';
 import { MultiSelect } from './variants/MultiSelect';
 import { SingleSelect } from './variants/SingleSelect';
@@ -12,31 +11,19 @@ import { SliderSelect } from './variants/SliderSelect';
 
 export const Select = <T extends string | string[]>({
   variant,
-  displayMode = SelectDisplayMode.Menu,
   ...props
 }: SelectProps<T>) => {
   if (variant === SelectVariant.Single) {
     return (
-      <SingleSelect
-        {...(props as unknown as SingleSelectProps<string>)}
-        displayMode={displayMode}
-      />
+      <SingleSelect {...(props as unknown as SingleSelectProps<string>)} />
     );
   }
 
   if (variant === SelectVariant.Slider) {
     return (
-      <SliderSelect
-        {...(props as unknown as SliderSelectProps<number[]>)}
-        displayMode={displayMode}
-      />
+      <SliderSelect {...(props as unknown as SliderSelectProps<number[]>)} />
     );
   }
 
-  return (
-    <MultiSelect
-      {...(props as unknown as MultiSelectProps<string[]>)}
-      displayMode={displayMode}
-    />
-  );
+  return <MultiSelect {...(props as unknown as MultiSelectProps<string[]>)} />;
 };
