+++ src/components/core/form/Select/Select.types.ts
@@ -13,11 +13,6 @@ export enum SelectSize {
   Medium = 'medium',
 }
 
-export enum SelectDisplayMode {
-  Menu = 'menu',
-  Drawer = 'drawer',
-}
-
 export interface SelectOption<T> {
   value: T;
   label: string;
@@ -26,7 +21,7 @@ export interface SelectOption<T> {
   sx?: SxProps<Theme>;
 }
 
-export interface BaseProps<T extends TData> {
+export interface SelectBaseProps<T extends TData> {
   options: SelectOption<T extends (infer U)[] ? U : T>[];
   value: T;
   onChange: (value: T) => void;
@@ -38,48 +33,26 @@ export interface BaseProps<T extends TData> {
   title?: string;
   required?: boolean;
   debounceMs?: number;
-  'data-testid'?: string;
-  displayMode?: SelectDisplayMode;
-}
-
-export interface MenuSelectProps<T extends TData> extends BaseProps<T> {
   menuPlacementX?: 'left' | 'right';
-  displayMode: SelectDisplayMode.Menu;
+  'data-testid'?: string;
 }
 
-export interface DrawerSelectProps<T extends TData> extends BaseProps<T> {
-  displayMode: SelectDisplayMode.Drawer;
-  showTrigger?: boolean;
-  open?: boolean;
-  onOpen?: () => void;
-  onClose?: () => void;
-  onBack?: () => void;
-}
-
-export type SelectBaseProps<T extends TData> =
-  | MenuSelectProps<T>
-  | DrawerSelectProps<T>;
-
-export type MultiSelectProps<T extends string[]> = SelectBaseProps<T> & {
+export interface MultiSelectProps<T extends string[]>
+  extends SelectBaseProps<T> {
   filterBy?: string;
   label: string;
-};
+}
 
-export type SingleSelectProps<T extends string> = SelectBaseProps<T>;
+export interface SingleSelectProps<T extends string>
+  extends SelectBaseProps<T> {}
 
-export type SliderSelectProps<T extends number[]> =
-  | (Omit<MenuSelectProps<T>, 'options'> & {
-      options: never[];
-      min: number;
-      max: number;
-      label: string;
-    })
-  | (Omit<DrawerSelectProps<T>, 'options'> & {
-      options: never[];
-      min: number;
-      max: number;
-      label: string;
-    });
+export interface SliderSelectProps<T extends number[]>
+  extends Omit<SelectBaseProps<T>, 'options'> {
+  options: never[];
+  min: number;
+  max: number;
+  label: string;
+}
 
 export type SelectProps<T extends TData> =
   | (MultiSelectProps<string[]> & { variant: SelectVariant.Multi })
