+++ src/components/core/form/Select/components/SelectBase.tsx
@@ -1,16 +1,17 @@
-import { SelectBaseProps, TData, SelectDisplayMode } from '../Select.types';
-import { PropsWithChildren } from 'react';
+import { SelectBaseProps, TData } from '../Select.types';
+import { PropsWithChildren, useCallback, useState } from 'react';
+import {
+  StyledSelect,
+  StyledSelectorContainer,
+  StyledSelectorContentContainer,
+  StyledMenuItem,
+  StyledMenuItemContentContainer,
+} from '../Select.styles';
+import KeyboardArrowDownRoundedIcon from '@mui/icons-material/KeyboardArrowDownRounded';
+import CheckIcon from '@mui/icons-material/Check';
 import { SelectProps } from '@mui/material/Select';
-import {
-  MenuDisplayMode,
-  MenuDisplayModeProps,
-} from '../displayMode/MenuDisplayMode';
-import {
-  DrawerDisplayMode,
-  DrawerDisplayModeProps,
-} from '../displayMode/DrawerDisplayMode';
-
-export { SelectDisplayMode as DisplayMode };
+import { SelectorLabel } from './SelectLabel';
+import Fade from '@mui/material/Fade';
 
 interface ExtendedSelectBaseProps<T extends TData>
   extends Omit<SelectBaseProps<T>, 'onChange'>,
@@ -18,18 +19,106 @@ interface ExtendedSelectBaseProps<T extends TData>
   multiple?: boolean;
   onChange: SelectProps['onChange'];
   selectorContent?: React.ReactNode;
-  title: string;
 }
 
 export const SelectBase = <T extends TData>({
-  displayMode = SelectDisplayMode.Menu,
+  children,
+  options,
+  value,
+  selectorContent,
+  menuPlacementX = 'left',
   ...rest
 }: ExtendedSelectBaseProps<T>) => {
-  if (displayMode === SelectDisplayMode.Drawer) {
-    return (
-      <DrawerDisplayMode {...(rest as unknown as DrawerDisplayModeProps<T>)} />
-    );
-  }
+  const [isOpen, setIsOpen] = useState(false);
 
-  return <MenuDisplayMode {...(rest as unknown as MenuDisplayModeProps<T>)} />;
+  const handleOpen = useCallback(() => {
+    setIsOpen(true);
+  }, []);
+
+  const handleClose = useCallback(() => {
+    setIsOpen(false);
+  }, []);
+
+  return (
+    <StyledSelect
+      {...rest}
+      value={value}
+      open={isOpen}
+      onOpen={handleOpen}
+      onClose={handleClose}
+      renderValue={() => (
+        <StyledSelectorContainer>
+          <StyledSelectorContentContainer>
+            {selectorContent}
+            <KeyboardArrowDownRoundedIcon
+              sx={{
+                height: 22,
+                width: 22,
+                transform: isOpen ? 'rotate(180deg)' : 'rotate(0deg)',
+                transition: 'transform 0.2s ease-in-out',
+              }}
+            />
+          </StyledSelectorContentContainer>
+        </StyledSelectorContainer>
+      )}
+      IconComponent={() => null}
+      autoWidth
+      variant="standard"
+      displayEmpty
+      MenuProps={{
+        anchorOrigin: {
+          vertical: 'bottom',
+          horizontal: menuPlacementX,
+        },
+        transformOrigin: {
+          vertical: 'top',
+          horizontal: menuPlacementX,
+        },
+        disableScrollLock: true,
+        slots: {
+          transition: Fade,
+        },
+        PaperProps: {
+          sx: (theme) => ({
+            backgroundColor: (theme.vars || theme).palette.surface1.main,
+            borderRadius: theme.spacing(3),
+            boxShadow: '0 4px 24px rgba(0, 0, 0, 0.08)',
+            marginTop: theme.spacing(1),
+            '& .MuiList-root': {
+              margin: theme.spacing(1),
+              padding: 0,
+              minWidth: '264px',
+              maxHeight: '300px',
+            },
+          }),
+        },
+      }}
+    >
+      {children}
+      {options.map((option) => (
+        <StyledMenuItem
+          disableRipple
+          key={option.value}
+          value={option.value}
+          sx={option.sx}
+        >
+          <StyledMenuItemContentContainer>
+            {option.icon}
+            <SelectorLabel label={option.label} />
+          </StyledMenuItemContentContainer>
+          {((Array.isArray(value) &&
+            (value as (string | number)[]).includes(option.value)) ||
+            (!Array.isArray(value) && value === option.value)) && (
+            <CheckIcon
+              sx={{
+                marginLeft: 'auto',
+                height: 16,
+                width: 16,
+              }}
+            />
+          )}
+        </StyledMenuItem>
+      ))}
+    </StyledSelect>
+  );
 };
