+++ src/components/core/form/Select/variants/SliderSelect.tsx
@@ -69,7 +69,6 @@ export const SliderSelect = <T extends number[]>({
     <SelectBase
       {...rest}
       value={''}
-      title={title ?? initialLabel}
       onChange={handleChange}
       multiple={false}
       selectorContent={
